+++ src/tile_physics.h
@@ -8,7 +8,7 @@ int tile_distance_right (ctrl_t *ctrl, moveable_t *obj);
 int tile_distance_up (ctrl_t *ctrl, moveable_t *obj);
 int tile_distance_down (ctrl_t *ctrl, moveable_t *obj);
 
-float clip_movement (float movement, int distance, float backup);
+double clip_movement (double movement, int distance, double backup);
 void tile_collisions_top (ctrl_t *ctrl, game_t *game, moveable_t *obj);
 void tile_collisions_bottom (ctrl_t *ctrl, game_t *game, moveable_t *obj);
